Traigo la shell nueva, pipes y cambios de kernel de main. Adapto los tests
nuestros a la firma nueva de sys_create_process (6 args) y los engancho en
la cmd_table. prio/sync quedan en la version del enunciado (prio incremental +
sync con waitpid); saco las viejas de tests.c. Compila con -Wall sin warnings.

+++ x64BareBones/Userland/SampleCodeModule/UserLib/include/stringLib.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint8_t strlen(const char *str);
-uint8_t strcmp(const char *s1, const char *s2);
+int strlen(const char *str);
+int strcmp(const char *s1, const char *s2);
 
 #endif
